requirements.txt
flask==3.0.3
twilio==9.3.*
python-dotenv==1.0.*
deep-translator==1.11.*
psycopg2-binary==2.9.*
gunicorn==23.0.*